fix makefile for make and make term
